+++ common/src/database/schema/sqlite/_008_add_princs_filter_fields_in_subscriptions_table.rs
@@ -1,0 +1,39 @@
+use anyhow::{anyhow, Result};
+use rusqlite::Connection;
+
+use crate::database::sqlite::SQLiteMigration;
+use crate::migration;
+
+pub(super) struct AddPrincsFilterFieldsInSubscriptionsTable;
+migration!(
+    AddPrincsFilterFieldsInSubscriptionsTable,
+    8,
+    "add princs_filter fields in subscriptions table"
+);
+
+impl SQLiteMigration for AddPrincsFilterFieldsInSubscriptionsTable {
+    fn up(&self, conn: &Connection) -> Result<()> {
+        conn.execute(
+            "ALTER TABLE subscriptions ADD COLUMN princs_filter_op TEXT",
+            [],
+        )
+        .map_err(|err| anyhow!("SQLiteError: {}", err))?;
+        conn.execute(
+            "ALTER TABLE subscriptions ADD COLUMN princs_filter_value TEXT",
+            [],
+        )
+        .map_err(|err| anyhow!("SQLiteError: {}", err))?;
+        Ok(())
+    }
+
+    fn down(&self, conn: &Connection) -> Result<()> {
+        conn.execute("ALTER TABLE subscriptions DROP COLUMN princs_filter_op", [])
+            .map_err(|err| anyhow!("SQLiteError: {}", err))?;
+        conn.execute(
+            "ALTER TABLE subscriptions DROP COLUMN princs_filter_value",
+            [],
+        )
+        .map_err(|err| anyhow!("SQLiteError: {}", err))?;
+        Ok(())
+    }
+}